refactor gtmachines into 2 new smaller classes for better readability
