+++ src/main/java/com/gregtechceu/gtceu/data/recipe/misc/MachineRecipeLoader.java
@@ -18,6 +18,7 @@ import com.gregtechceu.gtceu.common.data.GTItems;
 import com.gregtechceu.gtceu.common.data.GTMachines;
 import com.gregtechceu.gtceu.common.data.GTMaterials;
 import com.gregtechceu.gtceu.common.data.machines.GTAEMachines;
+import com.gregtechceu.gtceu.common.data.machines.GTMachineUtils;
 import com.gregtechceu.gtceu.config.ConfigHolder;
 import com.gregtechceu.gtceu.data.recipe.CustomTags;
 import com.gregtechceu.gtceu.data.recipe.VanillaRecipeHelper;
@@ -1420,7 +1421,7 @@ public class MachineRecipeLoader {
             }
         }
 
-        for (int tier : MULTI_HATCH_TIERS) {
+        for (int tier : GTMachineUtils.MULTI_HATCH_TIERS) {
             var tierName = VN[tier].toLowerCase();
 
             var importHatch4x = FLUID_IMPORT_HATCH_4X[tier];
@@ -1447,7 +1448,7 @@ public class MachineRecipeLoader {
                     'B', importHatch9x.asStack());
         }
 
-        for (int tier : DUAL_HATCH_TIERS) {
+        for (int tier : GTMachineUtils.DUAL_HATCH_TIERS) {
             var tierName = VN[tier].toLowerCase();
 
             var inputBuffer = DUAL_IMPORT_HATCH[tier];
