+++ src/main/java/com/gregtechceu/gtceu/integration/emi/multipage/MultiblockInfoEmiCategory.java
@@ -3,7 +3,7 @@ package com.gregtechceu.gtceu.integration.emi.multipage;
 import com.gregtechceu.gtceu.GTCEu;
 import com.gregtechceu.gtceu.api.machine.MultiblockMachineDefinition;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
-import com.gregtechceu.gtceu.common.data.GTMachines;
+import com.gregtechceu.gtceu.common.data.machines.GTMultiMachines;
 
 import net.minecraft.network.chat.Component;
 
@@ -16,7 +16,7 @@ public class MultiblockInfoEmiCategory extends EmiRecipeCategory {
     public static final MultiblockInfoEmiCategory CATEGORY = new MultiblockInfoEmiCategory();
 
     private MultiblockInfoEmiCategory() {
-        super(GTCEu.id("multiblock_info"), EmiStack.of(GTMachines.ELECTRIC_BLAST_FURNACE.getItem()));
+        super(GTCEu.id("multiblock_info"), EmiStack.of(GTMultiMachines.ELECTRIC_BLAST_FURNACE.getItem()));
     }
 
     public static void registerDisplays(EmiRegistry registry) {
